Reorganize settings routes
